# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This an example OTP implementation",
        "title": "dekamond example swagger API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
        "/check": {
            "post": {
                "description": "Accepts a phone number and an OTP code and return JWT token if they are valid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "login"
                ],
                "parameters": [
                    {
                        "description": "valid phone number and code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.CheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT containing user ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Accepts a phone number and create an OTP code if the phone number is valid and no OTP code is currently valid that number.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "valid phone number as string",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Retrieve users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "example": "09012345678",
                        "description": "A valid phone number for searching a specific user.",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2024-01-01,2025-10-12",
                        "description": "A date range to search for users who registered within that period in YYYY-MM-DD format, separated by a comma.",
                        "name": "register",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The page number of the results. Default is 1. Negative numbers and zero are treated as 1.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The number of items per page. Default is 10. Negative numbers and zero are treated as 1.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.SearchResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.CheckRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "123456"
                },
                "phone": {
                    "type": "string",
                    "example": "09012345678"
                }
            }
        },
        "app.LoginRequest": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "09012345678"
                }
            }
        },
        "app.SearchResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.User"
                    }
                }
            }
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "last_login": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "register_at": {
                    "type": "string"
                }
            }
        }
    }
}